Fix bugs across firmware, UI, and CI
- expr.rs: prevent i32 overflow in OP_LERP by promoting to i64
- protocol.js: remove incorrect COBS trailing-zero heuristic that
  could corrupt payloads ending with 0x00
- protocol.js: add bounds check in PostcardReader.u8() to throw on
  truncated data instead of silently returning undefined
- app.js: check readConfigFromUI() return value in withConfigMutation
  to avoid discarding unsaved DOM edits on expression compile errors
- app.js: await async Web Serial operations in cleanup() to prevent
  port state corruption on quick disconnect/reconnect
- input.rs: seed accelerometer EMA filter with first reading to avoid
  a CC message burst at startup when device is tilted
- config.rs: mark LDR pin as used in validate() for forward-compat
- serial.rs: add overflow flag to FrameAssembler so oversized frames
  are fully discarded instead of returned truncated
- release-firmware.yml: add --clobber to gh release upload for
  idempotent workflow re-runs
- expr.js: reject '>=' with a clear error message

+++ firmware/src/expr.rs
@@ -96,11 +96,11 @@ pub fn eval(program: &[u8; MAX_EXPR], len: u8, inputs: &ExprInputs, fallback: u8
                 val.clamp(lo, hi)
             }),
             OP_LERP => triop(&mut stack, &mut sp, |a, b, t| {
-                let a = i32::from(a);
-                let b = i32::from(b);
-                let t = i32::from(t);
+                let a = i64::from(a);
+                let b = i64::from(b);
+                let t = i64::from(t);
                 #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-                let val = (a + (b - a) * t / 127).clamp(0, i32::from(u16::MAX)) as u16;
+                let val = (a + (b - a) * t / 127).clamp(0, i64::from(u16::MAX)) as u16;
                 val
             }),
             OP_SCALE => {
